restructure project and simplify main entry point
